fix: address XFlickr30k-Co review feedback
